added default-enrollment-task Policy Set, refactored policy set base classes
