+++ rinad/src/ipcp/enrollment-task.cc
@@ -37,12 +37,6 @@
 
 namespace rinad {
 
-//	CLASS EnrollmentInformationRequest
-EnrollmentInformationRequest::EnrollmentInformationRequest() {
-	address_ = 0;
-	allowed_to_start_early_ = false;
-}
-
 //Class WatchdogTimerTask
 WatchdogTimerTask::WatchdogTimerTask(WatchdogRIBObject * watchdog,
                                      rina::Timer *       timer,
@@ -218,793 +212,6 @@ std::string AddressRIBObject::get_displayable_value() {
     return ss.str();
 }
 
-//Class BaseEnrollmentStateMachine
-const std::string BaseEnrollmentStateMachine::CONNECT_RESPONSE_TIMEOUT =
-		"Timeout waiting for connect response";
-const std::string BaseEnrollmentStateMachine::START_RESPONSE_TIMEOUT =
-		"Timeout waiting for start response";
-const std::string BaseEnrollmentStateMachine::START_IN_BAD_STATE =
-		"Received a START message in a wrong state";
-const std::string BaseEnrollmentStateMachine::STOP_ENROLLMENT_TIMEOUT =
-		"Timeout waiting for stop enrolment response";
-const std::string BaseEnrollmentStateMachine::STOP_IN_BAD_STATE =
-		"Received a STOP message in a wrong state";
-const std::string BaseEnrollmentStateMachine::STOP_WITH_NO_OBJECT_VALUE =
-		"Received STOP message with null object value";
-const std::string BaseEnrollmentStateMachine::READ_RESPONSE_TIMEOUT =
-		"Timeout waiting for read response";
-const std::string BaseEnrollmentStateMachine::PROBLEMS_COMMITTING_ENROLLMENT_INFO =
-		"Problems commiting enrollment information";
-const std::string BaseEnrollmentStateMachine::START_TIMEOUT =
-		"Timeout waiting for start";
-const std::string BaseEnrollmentStateMachine::READ_RESPONSE_IN_BAD_STATE =
-		"Received a READ_RESPONSE message in a wrong state";
-const std::string BaseEnrollmentStateMachine::UNSUCCESSFULL_READ_RESPONSE =
-		"Received an unsuccessful read response or a read response with a null object value";
-const std::string BaseEnrollmentStateMachine::UNSUCCESSFULL_START =
-		"Received unsuccessful start request";
-const std::string BaseEnrollmentStateMachine::CONNECT_IN_NOT_NULL =
-		"Received a CONNECT message while not in NULL state";
-const std::string BaseEnrollmentStateMachine::ENROLLMENT_NOT_ALLOWED =
-		"Enrollment rejected by security manager";
-const std::string BaseEnrollmentStateMachine::START_ENROLLMENT_TIMEOUT =
-		"Timeout waiting for start enrollment request";
-const std::string BaseEnrollmentStateMachine::STOP_ENROLLMENT_RESPONSE_TIMEOUT =
-		"Timeout waiting for stop enrollment response";
-const std::string BaseEnrollmentStateMachine::STOP_RESPONSE_IN_BAD_STATE =
-		"Received a STOP Response message in a wrong state";
-
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_CONNECT_RESPONSE =
-		"WAIT_CONNECT_RESPONSE";
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_START_ENROLLMENT_RESPONSE =
-		"WAIT_START_ENROLLMENT_RESPONSE";
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_READ_RESPONSE =
-		"WAIT_READ_RESPONSE";
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_START =
-		"WAIT_START";
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_START_ENROLLMENT =
-		"WAIT_START_ENROLLMENT";
-const std::string BaseEnrollmentStateMachine::STATE_WAIT_STOP_ENROLLMENT_RESPONSE =
-		"WAIT_STOP_ENROLLMENT_RESPONSE";
-
-BaseEnrollmentStateMachine::BaseEnrollmentStateMachine(IPCProcess * ipc_process,
-			const rina::ApplicationProcessNamingInformation& remote_naming_info,
-			int timeout, rina::ApplicationProcessNamingInformation * supporting_dif_name) :
-				rina::IEnrollmentStateMachine(ipc_process, true, remote_naming_info,
-						timeout, supporting_dif_name)
-{
-	ipc_process_ = ipc_process;
-	cdap_session_manager_ = ipc_process->cdap_session_manager_;
-	encoder_ = ipc_process->encoder_;
-}
-
-void BaseEnrollmentStateMachine::sendDIFDynamicInformation() {
-	//Send DirectoryForwardingTableEntries
-	sendDFTEntries();
-
-	//Send neighbors (including myself)
-	sendNeighbors();
-}
-
-void BaseEnrollmentStateMachine::sendDFTEntries() {
-	rina::BaseRIBObject * dftEntrySet;
-	std::list<rina::BaseRIBObject *>::const_iterator it;
-	std::list<rina::DirectoryForwardingTableEntry *> dftEntries;
-
-	try {
-		dftEntrySet = rib_daemon_->readObject(EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_CLASS,
-				EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_NAME);
-		for (it = dftEntrySet->get_children().begin();
-				it != dftEntrySet->get_children().end(); ++it) {
-			dftEntries.push_back((rina::DirectoryForwardingTableEntry*) (*it)->get_value());
-		}
-
-		if (dftEntries.size() == 0) {
-			LOG_IPCP_DBG("No DFT entries to be sent");
-			return;
-		}
-
-		rina::RIBObjectValue robject_value;
-		robject_value.type_ = rina::RIBObjectValue::complextype;
-		robject_value.complex_value_ = &dftEntries;
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		rib_daemon_->remoteCreateObject(EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_CLASS,
-				EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_NAME, robject_value, 0, remote_id, 0);
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems sending DFT entries: %s", e.what());
-	}
-}
-
-// Class EnrolleeStateMachine
-EnrolleeStateMachine::EnrolleeStateMachine(IPCProcess * ipc_process,
-		const rina::ApplicationProcessNamingInformation& remote_naming_info,
-		int timeout): BaseEnrollmentStateMachine(ipc_process, remote_naming_info,
-				timeout, 0) {
-	was_dif_member_before_enrollment_ = false;
-	enrollment_request_ = 0;
-	last_scheduled_task_ = 0;
-	allowed_to_start_early_ = false;
-	stop_request_invoke_id_ = 0;
-}
-
-EnrolleeStateMachine::~EnrolleeStateMachine() {
-}
-
-void EnrolleeStateMachine::initiateEnrollment(rina::EnrollmentRequest * enrollmentRequest,
-					      int portId)
-{
-	rina::ScopedLock g(*lock_);
-
-	enrollment_request_ = enrollmentRequest;
-	remote_peer_->address_ = enrollment_request_->neighbor_->address_;
-	remote_peer_->name_ = enrollment_request_->neighbor_->name_;
-	remote_peer_->supporting_dif_name_ = enrollment_request_->neighbor_->supporting_dif_name_;
-	remote_peer_->underlying_port_id_ = enrollment_request_->neighbor_->underlying_port_id_;
-	remote_peer_->supporting_difs_ = enrollment_request_->neighbor_->supporting_difs_;
-
-	if (state_ != STATE_NULL) {
-		throw rina::Exception("Enrollee state machine not in NULL state");
-	}
-
-	try{
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = portId;
-
-		rib_daemon_->openApplicationConnection(rina::CDAPMessage::AUTH_NONE, rina::AuthValue(), "", IPCProcess::MANAGEMENT_AE,
-				remote_peer_->name_.processInstance, remote_peer_->name_.processName, "",
-				IPCProcess::MANAGEMENT_AE, ipc_process_->get_instance(),
-				ipc_process_->get_name(), remote_id);
-
-		port_id_ = portId;
-
-		//Set timer
-		last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this, CONNECT_RESPONSE_TIMEOUT, true);
-		timer_->scheduleTask(last_scheduled_task_, timeout_);
-
-		//Update state
-		state_ = STATE_WAIT_CONNECT_RESPONSE;
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending M_CONNECT message: %s", e.what());
-		abortEnrollment(remote_peer_->name_, port_id_, std::string(e.what()), true, false);
-	}
-}
-
-void EnrolleeStateMachine::connectResponse(int result,
-		const std::string& result_reason) {
-	rina::ScopedLock g(*lock_);
-
-	if (state_ != STATE_WAIT_CONNECT_RESPONSE) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				"Message received in wrong order", true, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-	if (result != 0) {
-		state_ = STATE_NULL;
-		enrollment_task_->enrollmentFailed(remote_peer_->get_name(), port_id_,
-				result_reason, true, true);
-		return;
-	}
-
-	//Send M_START with EnrollmentInformation object
-	try{
-		EnrollmentInformationRequest eiRequest;
-		std::list<rina::ApplicationProcessNamingInformation>::const_iterator it;
-		std::vector<rina::ApplicationRegistration *> registrations =
-				rina::extendedIPCManager->getRegisteredApplications();
-		for(unsigned int i=0; i<registrations.size(); i++) {
-			for(it = registrations[i]->DIFNames.begin();
-					it != registrations[i]->DIFNames.end(); ++it) {;
-				eiRequest.supporting_difs_.push_back(*it);
-			}
-		}
-
-		if (ipc_process_->get_address() != 0) {
-			was_dif_member_before_enrollment_ = true;
-			eiRequest.address_ = ipc_process_->get_address();
-		} else {
-			rina::DIFInformation difInformation;
-			difInformation.dif_name_ = enrollment_request_->event_.dafName;
-			ipc_process_->set_dif_information(difInformation);
-		}
-
-		rina::RIBObjectValue object_value;
-		object_value.type_ = rina::RIBObjectValue::complextype;
-		object_value.complex_value_ = &eiRequest;
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		rib_daemon_->remoteStartObject(EncoderConstants::ENROLLMENT_INFO_OBJECT_CLASS,
-				EncoderConstants::ENROLLMENT_INFO_OBJECT_NAME, object_value, 0, remote_id, this);
-
-		LOG_IPCP_DBG("Sent a M_START Message to portid: %d", port_id_);
-
-		//Set timer
-		last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this, START_RESPONSE_TIMEOUT, true);
-		timer_->scheduleTask(last_scheduled_task_, timeout_);
-
-		//Update state
-		state_ = STATE_WAIT_START_ENROLLMENT_RESPONSE;
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending M_START request message: %s", e.what());
-		//TODO what to do?
-	}
-}
-
-void EnrolleeStateMachine::startResponse(int result, const std::string& result_reason,
-		void * object_value, rina::CDAPSessionDescriptor * session_descriptor) {
-	rina::ScopedLock g(*lock_);
-
-	if (!isValidPortId(session_descriptor)){
-		return;
-	}
-
-	if (state_ != STATE_WAIT_START_ENROLLMENT_RESPONSE) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				START_IN_BAD_STATE, true, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-	if (result != 0) {
-		state_ = STATE_NULL;
-		enrollment_task_->enrollmentFailed(remote_peer_->get_name(), port_id_,
-				result_reason, true, true);
-		return;
-	}
-
-	//Update address
-	if (object_value) {
-		EnrollmentInformationRequest * response =
-				(EnrollmentInformationRequest *) object_value;
-
-		unsigned int address = response->address_;
-		delete response;
-
-		try {
-			rib_daemon_->writeObject(EncoderConstants::ADDRESS_RIB_OBJECT_CLASS,
-					EncoderConstants::ADDRESS_RIB_OBJECT_NAME, &address);
-		} catch (rina::Exception &e) {
-			LOG_IPCP_ERR("Problems writing RIB object: %s", e.what());
-		}
-	}
-
-	//Set timer
-	last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this, STOP_ENROLLMENT_TIMEOUT, true);
-	timer_->scheduleTask(last_scheduled_task_, timeout_);
-
-	//Update state
-	state_ = STATE_WAIT_STOP_ENROLLMENT_RESPONSE;
-}
-
-void EnrolleeStateMachine::stop(EnrollmentInformationRequest * eiRequest, int invoke_id,
-		rina::CDAPSessionDescriptor * cdapSessionDescriptor) {
-	rina::ScopedLock g(*lock_);
-
-	if (!isValidPortId(cdapSessionDescriptor)){
-		return;
-	}
-
-	if (state_ != STATE_WAIT_STOP_ENROLLMENT_RESPONSE) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				STOP_IN_BAD_STATE, true, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-	//Check if I'm allowed to start early
-	if (!eiRequest->allowed_to_start_early_){
-		abortEnrollment(remote_peer_->name_, port_id_, STOP_WITH_NO_OBJECT_VALUE, true, true);
-		return;
-	}
-
-	allowed_to_start_early_ = eiRequest->allowed_to_start_early_;
-	stop_request_invoke_id_ = invoke_id;
-
-	//If the enrollee is also a member of the DIF, send dynamic information to the enroller as well
-	if (ipc_process_->get_operational_state() == ASSIGNED_TO_DIF) {
-		sendDFTEntries();
-	}
-
-	//Request more information or start
-	try{
-		requestMoreInformationOrStart();
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems requesting more information or starting: %s", e.what());
-		abortEnrollment(remote_peer_->name_, port_id_, std::string(e.what()), true, true);
-	}
-}
-
-void EnrolleeStateMachine::requestMoreInformationOrStart() {
-	if (sendNextObjectRequired()){
-		//Set timer
-		last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this,
-				READ_RESPONSE_TIMEOUT, true);
-		timer_->scheduleTask(last_scheduled_task_, timeout_);
-
-		//Update state
-		state_ = STATE_WAIT_READ_RESPONSE;
-		return;
-	}
-
-	//No more information is required, if I'm allowed to start early,
-	//commit the enrollment information, set operational status to true
-	//and send M_STOP_R. If not, just send M_STOP_R
-	rina::RIBObjectValue object_value;
-	rina::RemoteProcessId remote_id;
-	remote_id.port_id_ = port_id_;
-
-	if (allowed_to_start_early_){
-		try{
-			commitEnrollment();
-
-			rib_daemon_->remoteStopObjectResponse("", "", object_value, 0, "", stop_request_invoke_id_, remote_id);
-
-			enrollmentCompleted();
-		}catch(rina::Exception &e){
-			LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-
-			rib_daemon_->remoteStopObjectResponse("", "", object_value, -1,
-					PROBLEMS_COMMITTING_ENROLLMENT_INFO, stop_request_invoke_id_, remote_id);
-
-			abortEnrollment(remote_peer_->name_, port_id_, PROBLEMS_COMMITTING_ENROLLMENT_INFO, true, true);
-		}
-
-		return;
-	}
-
-	try {
-		rib_daemon_->remoteStopObjectResponse("", "", object_value, 0, "", stop_request_invoke_id_, remote_id);
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-	}
-
-	last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this, START_TIMEOUT, true);
-	timer_->scheduleTask(last_scheduled_task_, timeout_);
-	state_ = STATE_WAIT_START;
-}
-
-bool EnrolleeStateMachine::sendNextObjectRequired() {
-	bool result = false;
-	rina::DIFInformation difInformation = ipc_process_->get_dif_information();
-
-	rina::RemoteProcessId remote_id;
-	remote_id.port_id_ = port_id_;
-	std::string object_class;
-	std::string object_name;
-	if (!difInformation.dif_configuration_.efcp_configuration_.data_transfer_constants_.isInitialized()) {
-		object_class = EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_CLASS;
-		object_name = EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_CLASS;
-		result = true;
-	} else if (difInformation.dif_configuration_.efcp_configuration_.qos_cubes_.size() == 0){
-		object_class = EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_CLASS;
-		object_name = EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_NAME;
-		result = true;
-	}else if (ipc_process_->get_neighbors().size() == 0){
-		object_class = rina::NeighborSetRIBObject::NEIGHBOR_SET_RIB_OBJECT_CLASS;
-		object_name = rina::NeighborSetRIBObject::NEIGHBOR_SET_RIB_OBJECT_NAME;
-		result = true;
-	}
-
-	if (result) {
-		try{
-			rib_daemon_->remoteReadObject(object_class, object_name, 0, remote_id, this);
-		} catch (rina::Exception &e) {
-			LOG_IPCP_WARN("Problems executing remote operation: %s", e.what());
-		}
-	}
-
-	return result;
-}
-
-void EnrolleeStateMachine::commitEnrollment() {
-	try {
-		rib_daemon_->startObject(EncoderConstants::OPERATIONAL_STATUS_RIB_OBJECT_CLASS,
-				EncoderConstants::OPERATIONAL_STATUS_RIB_OBJECT_NAME, 0);
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems starting RIB object: %s", e.what());
-	}
-}
-
-void EnrolleeStateMachine::enrollmentCompleted() {
-	delete timer_;
-	timer_ = 0;
-	state_ = STATE_ENROLLED;
-
-	//Create or update the neighbor information in the RIB
-	createOrUpdateNeighborInformation(true);
-
-	//Send DirectoryForwardingTableEntries
-	sendCreateInformation(EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_CLASS,
-			EncoderConstants::DFT_ENTRY_SET_RIB_OBJECT_NAME);
-
-	enrollment_task_->enrollmentCompleted(remote_peer_, true);
-
-	//Notify the kernel
-	if (!was_dif_member_before_enrollment_) {
-		try {
-			rina::kernelIPCProcess->assignToDIF(ipc_process_->get_dif_information());
-		} catch(rina::Exception &e) {
-			LOG_IPCP_ERR("Problems communicating with the Kernel components of the IPC Processs: %s",
-					e.what());
-		}
-	}
-
-	//Notify the IPC Manager
-	if (enrollment_request_){
-		try {
-			std::list<rina::Neighbor> neighbors;
-			neighbors.push_back(*remote_peer_);
-			rina::extendedIPCManager->enrollToDIFResponse(enrollment_request_->event_,
-					0, neighbors, ipc_process_->get_dif_information());
-		} catch (rina::Exception &e) {
-			LOG_IPCP_ERR("Problems sending message to IPC Manager: %s", e.what());
-		}
-	}
-
-	LOG_IPCP_INFO("Remote IPC Process enrolled!");
-
-	delete enrollment_request_;
-}
-
-void EnrolleeStateMachine::readResponse(int result, const std::string& result_reason,
-		void * object_value, const std::string& object_name,
-		rina::CDAPSessionDescriptor * session_descriptor) {
-	rina::ScopedLock g(*lock_);
-
-	if (!isValidPortId(session_descriptor)){
-		return;
-	}
-
-	if (state_ != STATE_WAIT_READ_RESPONSE) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				READ_RESPONSE_IN_BAD_STATE, true, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-
-	if (result != 0 || object_value == 0){
-		abortEnrollment(remote_peer_->name_, port_id_,
-				result_reason, true, true);
-		return;
-	}
-
-	if (object_name.compare(EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_NAME) == 0){
-		try{
-			rina::DataTransferConstants * constants =
-					(rina::DataTransferConstants *) object_value;
-			rib_daemon_->createObject(EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_CLASS,
-					EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_NAME, constants, 0);
-		}catch(rina::Exception &e){
-			LOG_IPCP_ERR("Problems creating RIB object: %s", e.what());
-		}
-	}else if (object_name.compare(EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_NAME) == 0){
-		try{
-			std::list<rina::QoSCube *> * cubes =
-					(std::list<rina::QoSCube *> *) object_value;
-			rib_daemon_->createObject(EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_CLASS,
-					EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_NAME, cubes, 0);
-		}catch(rina::Exception &e){
-			LOG_IPCP_ERR("Problems creating RIB object: %s", e.what());
-		}
-	}else if (object_name.compare(rina::NeighborSetRIBObject::NEIGHBOR_SET_RIB_OBJECT_NAME) == 0){
-		try{
-			std::list<rina::Neighbor *> * neighbors =
-					(std::list<rina::Neighbor *> *) object_value;
-			rib_daemon_->createObject(rina::NeighborSetRIBObject::NEIGHBOR_SET_RIB_OBJECT_CLASS,
-					rina::NeighborSetRIBObject::NEIGHBOR_SET_RIB_OBJECT_NAME,
-					neighbors, 0);
-		}catch(rina::Exception &e){
-			LOG_IPCP_ERR("Problems creating RIB object: %s", e.what());
-		}
-	}else{
-		LOG_IPCP_WARN("The object to be created is not required for enrollment");
-	}
-
-	//Request more information or proceed with the enrollment program
-	requestMoreInformationOrStart();
-}
-
-void EnrolleeStateMachine::start(int result, const std::string& result_reason,
-		rina::CDAPSessionDescriptor * session_descriptor) {
-	rina::ScopedLock g(*lock_);
-
-	if (!isValidPortId(session_descriptor)){
-		return;
-	}
-
-	if (state_ == STATE_ENROLLED) {
-		return;
-	}
-
-	if (state_ != STATE_WAIT_START) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				START_IN_BAD_STATE, true, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-
-	if (result != 0){
-		abortEnrollment(remote_peer_->name_, port_id_,
-				result_reason, true, true);
-		return;
-	}
-
-	try{
-		commitEnrollment();
-		enrollmentCompleted();
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems commiting enrollment: %s", e.what());
-		abortEnrollment(remote_peer_->name_, port_id_,
-				PROBLEMS_COMMITTING_ENROLLMENT_INFO, true, true);
-	}
-}
-
-//Class EnrollerStateMachine
-EnrollerStateMachine::EnrollerStateMachine(IPCProcess * ipc_process,
-		const rina::ApplicationProcessNamingInformation& remote_naming_info, int timeout,
-		rina::ApplicationProcessNamingInformation * supporting_dif_name):
-		BaseEnrollmentStateMachine(ipc_process, remote_naming_info,
-					timeout, supporting_dif_name){
-	security_manager_ = ipc_process->security_manager_;
-	namespace_manager_ = ipc_process->namespace_manager_;
-	enroller_ = true;
-}
-
-EnrollerStateMachine::~EnrollerStateMachine() {
-}
-
-void EnrollerStateMachine::connect(int invoke_id, rina::CDAPSessionDescriptor * session_descriptor) {
-	rina::ScopedLock g(*lock_);
-        ISecurityManagerPs *smps = dynamic_cast<ISecurityManagerPs *>(security_manager_->ps);
-
-        assert(smps);
-
-	if (state_ != STATE_NULL) {
-		abortEnrollment(remote_peer_->name_, session_descriptor->port_id_,
-				CONNECT_IN_NOT_NULL, false, true);
-		return;
-	}
-
-	LOG_IPCP_DBG("Authenticating IPC process %s-%s ...", session_descriptor->dest_ap_name_.c_str(),
-			session_descriptor->dest_ap_inst_.c_str());
-	remote_peer_->name_.processName = session_descriptor->dest_ap_name_;
-	remote_peer_->name_.processInstance = session_descriptor->dest_ap_inst_;
-
-	//TODO Authenticate sender
-	LOG_IPCP_DBG("Authentication successful, deciding if new member can join the DIF...");
-	if (!smps->isAllowedToJoinDIF(*remote_peer_)) {
-		LOG_IPCP_WARN("Security Manager rejected enrollment attempt, aborting enrollment");
-		abortEnrollment(remote_peer_->name_, port_id_,
-				ENROLLMENT_NOT_ALLOWED, false, true);
-		return;
-	}
-
-
-	//Send M_CONNECT_R
-	port_id_ = session_descriptor->port_id_;
-	try{
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		rib_daemon_->openApplicationConnectionResponse(rina::CDAPMessage::AUTH_NONE,
-				rina::AuthValue(), session_descriptor->dest_ae_inst_, IPCProcess::MANAGEMENT_AE,
-				session_descriptor->dest_ap_inst_, session_descriptor->dest_ap_name_, 0, "", session_descriptor->src_ae_inst_,
-				IPCProcess::MANAGEMENT_AE, session_descriptor->src_ap_inst_, session_descriptor->src_ap_name_,
-				invoke_id, remote_id);
-
-		//Set timer
-		last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this, START_ENROLLMENT_TIMEOUT, true);
-		timer_->scheduleTask(last_scheduled_task_, timeout_);
-		LOG_IPCP_DBG("M_CONNECT_R sent to portID %d. Waiting for start enrollment request message", port_id_);
-
-		state_ = STATE_WAIT_START_ENROLLMENT;
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-		abortEnrollment(remote_peer_->name_, port_id_,
-						std::string(e.what()), false, true);
-	}
-}
-
-void EnrollerStateMachine::sendNegativeStartResponseAndAbortEnrollment(int result, const std::string&
-		resultReason, int invoke_id) {
-	try{
-		rina::RIBObjectValue robject_value;
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		rib_daemon_->remoteStartObjectResponse("", "", robject_value, result,
-				resultReason, invoke_id, remote_id);
-
-		abortEnrollment(remote_peer_->name_, port_id_, resultReason, false, true);
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-	}
-}
-
-void EnrollerStateMachine::sendDIFStaticInformation() {
-	sendCreateInformation(EncoderConstants::WHATEVERCAST_NAME_SET_RIB_OBJECT_CLASS,
-			EncoderConstants::WHATEVERCAST_NAME_SET_RIB_OBJECT_NAME);
-
-	sendCreateInformation(EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_CLASS,
-			EncoderConstants::DATA_TRANSFER_CONSTANTS_RIB_OBJECT_NAME);
-
-	sendCreateInformation(EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_CLASS,
-			EncoderConstants::QOS_CUBE_SET_RIB_OBJECT_NAME);
-}
-
-void EnrollerStateMachine::start(EnrollmentInformationRequest * eiRequest, int invoke_id,
-		rina::CDAPSessionDescriptor * cdapSessionDescriptor) {
-	rina::ScopedLock g(*lock_);
-
-    INamespaceManagerPs *nsmps = dynamic_cast<INamespaceManagerPs *>(namespace_manager_->ps);
-    assert(nsmps);
-
-	if (!isValidPortId(cdapSessionDescriptor)){
-		return;
-	}
-
-	if (state_ != STATE_WAIT_START_ENROLLMENT) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				START_IN_BAD_STATE, false, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-
-	bool requiresInitialization = false;
-
-	LOG_IPCP_DBG("Remote IPC Process address: %u", eiRequest->address_);
-
-	if (!eiRequest) {
-		requiresInitialization = true;
-	} else {
-		try {
-			if (!nsmps->isValidAddress(eiRequest->address_, remote_peer_->name_.processName,
-					remote_peer_->name_.processInstance)) {
-				requiresInitialization = true;
-			}
-
-			std::list<rina::ApplicationProcessNamingInformation>::const_iterator it;
-			for (it = eiRequest->supporting_difs_.begin();
-					it != eiRequest->supporting_difs_.end(); ++it) {
-				remote_peer_->supporting_difs_.push_back(*it);
-			}
-		}catch (rina::Exception &e) {
-			LOG_IPCP_ERR("%s", e.what());
-			sendNegativeStartResponseAndAbortEnrollment(-1, std::string(e.what()), invoke_id);
-			return;
-		}
-	}
-
-	if (requiresInitialization){
-		unsigned int address = nsmps->getValidAddress(remote_peer_->name_.processName,
-				remote_peer_->name_.processInstance);
-
-		if (address == 0){
-			sendNegativeStartResponseAndAbortEnrollment(-1, "Could not assign a valid address", invoke_id);
-			return;
-		}
-
-		LOG_IPCP_DBG("Remote IPC Process requires initialization, assigning address %u", address);
-		eiRequest->address_ = address;
-	}
-
-	try {
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-		rina::RIBObjectValue object_value;
-
-		if (requiresInitialization) {
-			object_value.type_ = rina::RIBObjectValue::complextype;
-			object_value.complex_value_ = eiRequest;
-		}
-
-		rib_daemon_->remoteStartObjectResponse(EncoderConstants::ENROLLMENT_INFO_OBJECT_CLASS,
-				EncoderConstants::ENROLLMENT_INFO_OBJECT_NAME, object_value, 0, "",
-				invoke_id, remote_id);
-
-		remote_peer_->address_ = eiRequest->address_;
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-		delete eiRequest;
-		sendNegativeStartResponseAndAbortEnrollment(-1, std::string(e.what()), invoke_id);
-		return;
-	}
-
-	//If initialization is required send the M_CREATEs
-	if (requiresInitialization){
-		sendDIFStaticInformation();
-	}
-
-	sendDIFDynamicInformation();
-
-	//Send the M_STOP request
-	try {
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		eiRequest->allowed_to_start_early_ = true;
-		rina::RIBObjectValue object_value;
-		object_value.type_ = rina::RIBObjectValue::complextype;
-		object_value.complex_value_ = eiRequest;
-
-		rib_daemon_->remoteStopObject(EncoderConstants::ENROLLMENT_INFO_OBJECT_CLASS,
-				EncoderConstants::ENROLLMENT_INFO_OBJECT_NAME, object_value, 0, remote_id, this);
-
-		delete eiRequest;
-	} catch(rina::Exception &e) {
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-		delete eiRequest;
-		sendNegativeStartResponseAndAbortEnrollment(-1, std::string(e.what()), invoke_id);
-		return;
-	}
-
-	//Set timer
-	last_scheduled_task_ = new rina::EnrollmentFailedTimerTask(this,
-			STOP_ENROLLMENT_RESPONSE_TIMEOUT, true);
-	timer_->scheduleTask(last_scheduled_task_, timeout_);
-
-	LOG_IPCP_DBG("Waiting for stop enrollment response message");
-	state_ = STATE_WAIT_STOP_ENROLLMENT_RESPONSE;
-}
-
-void EnrollerStateMachine::stopResponse(int result, const std::string& result_reason,
-		void * object_value, rina::CDAPSessionDescriptor * session_descriptor) {
-	rina::ScopedLock g(*lock_);
-
-	(void) object_value;
-
-	if (!isValidPortId(session_descriptor)){
-		return;
-	}
-
-	if (state_ != STATE_WAIT_STOP_ENROLLMENT_RESPONSE) {
-		abortEnrollment(remote_peer_->name_, port_id_,
-				STOP_RESPONSE_IN_BAD_STATE, false, true);
-		return;
-	}
-
-	timer_->cancelTask(last_scheduled_task_);
-	if (result != 0){
-		state_ = STATE_NULL;
-		enrollment_task_->enrollmentFailed(remote_peer_->name_, port_id_,
-				result_reason, false, true);
-		return;
-	}
-
-	try{
-		rina::RIBObjectValue robject_value;
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = port_id_;
-
-		rib_daemon_->remoteStartObject(EncoderConstants::OPERATIONAL_STATUS_RIB_OBJECT_CLASS,
-				EncoderConstants::OPERATIONAL_STATUS_RIB_OBJECT_NAME, robject_value, 0, remote_id, 0);
-	} catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems sending CDAP Message: %s", e.what());
-	}
-
-	enrollmentCompleted();
-}
-
-void EnrollerStateMachine::enrollmentCompleted() {
-	delete timer_;
-	timer_ = 0;
-
-	state_ = STATE_ENROLLED;
-
-	createOrUpdateNeighborInformation(true);
-
-	enrollment_task_->enrollmentCompleted(remote_peer_, false);
-
-	LOG_IPCP_INFO("Remote IPC Process enrolled!");
-}
-
 //Main function of the Neighbor Enroller thread
 void * doNeighborsEnrollerWork(void * arg) {
 	IPCProcess * ipcProcess = (IPCProcess *) arg;
@@ -1086,8 +293,6 @@ void EnrollmentTask::populateRIB()
 		rina::BaseRIBObject * ribObject = new rina::NeighborSetRIBObject(ipcp,
 									         ipcp->rib_daemon_);
 		rib_daemon_->addRIBObject(ribObject);
-		ribObject = new EnrollmentRIBObject(ipcp);
-		rib_daemon_->addRIBObject(ribObject);
 		ribObject = new OperationalStatusRIBObject(ipcp);
 		rib_daemon_->addRIBObject(ribObject);
 		ribObject = new AddressRIBObject(ipcp);
@@ -1115,6 +320,11 @@ void EnrollmentTask::set_dif_configuration(const rina::DIFConfiguration& dif_con
 	neighbors_enroller_ = new rina::Thread(threadAttributes,
 			&doNeighborsEnrollerWork, (void *) ipcp);
 	LOG_IPCP_DBG("Started Neighbors enroller thread");
+
+	//Apply configuration to policy set
+	IPCPEnrollmentTaskPS * ipcp_ps = dynamic_cast<IPCPEnrollmentTaskPS *>(ps);
+	assert(ipcp_ps);
+	ipcp_ps->set_dif_configuration(dif_configuration);
 }
 
 void EnrollmentTask::processEnrollmentRequestEvent(rina::EnrollToDAFRequestEvent* event)
@@ -1202,37 +412,6 @@ void EnrollmentTask::initiateEnrollment(rina::EnrollmentRequest * request)
 	port_ids_pending_to_be_allocated_.put(handle, request);
 }
 
-rina::IEnrollmentStateMachine * EnrollmentTask::createEnrollmentStateMachine(
-		const rina::ApplicationProcessNamingInformation& apNamingInfo, int portId,
-		bool enrollee, const rina::ApplicationProcessNamingInformation& supportingDifName)
-{
-	BaseEnrollmentStateMachine * stateMachine = 0;
-
-	if (apNamingInfo.entityName.compare("") == 0 ||
-			apNamingInfo.entityName.compare(IPCProcess::MANAGEMENT_AE) == 0) {
-		if (enrollee){
-			stateMachine = new EnrolleeStateMachine(ipcp,
-					apNamingInfo, timeout_);
-		}else{
-			rina::ApplicationProcessNamingInformation * sdname =
-					new rina::ApplicationProcessNamingInformation(supportingDifName.processName,
-							supportingDifName.processInstance);
-			stateMachine = new EnrollerStateMachine(ipcp,
-					apNamingInfo, timeout_, sdname);
-		}
-
-		std::stringstream ss;
-		ss<<apNamingInfo.processName<<"-"<<portId;
-		state_machines_.put(ss.str(), stateMachine);
-
-		LOG_IPCP_DBG("Created a new Enrollment state machine for remote IPC process: %s",
-				apNamingInfo.getEncodedString().c_str());
-		return stateMachine;
-	}
-
-	throw rina::Exception("Unknown application entity for enrollment");
-}
-
 void EnrollmentTask::connect(int invoke_id,
 		rina::CDAPSessionDescriptor * session_descriptor)
 {
@@ -1255,11 +434,20 @@ void EnrollmentTask::connect(int invoke_id,
 			rina::RemoteProcessId remote_id;
 			remote_id.port_id_ = session_descriptor->port_id_;
 
-			rib_daemon_->openApplicationConnectionResponse(rina::CDAPMessage::AUTH_NONE,
-					rina::AuthValue(), session_descriptor->dest_ae_inst_, session_descriptor->dest_ae_name_,
-					session_descriptor->dest_ap_inst_, session_descriptor->dest_ap_name_, -2, message,
-					session_descriptor->src_ae_inst_, session_descriptor->src_ae_name_,
-					session_descriptor->src_ap_inst_, session_descriptor->src_ap_name_, invoke_id, remote_id);
+			rib_daemon_->openApplicationConnectionResponse(session_descriptor->auth_mech_,
+								       rina::AuthValue(),
+								       session_descriptor->dest_ae_inst_,
+								       session_descriptor->dest_ae_name_,
+								       session_descriptor->dest_ap_inst_,
+								       session_descriptor->dest_ap_name_,
+								       rina::CDAPErrorCodes::CONNECTION_REJECTED_ERROR,
+								       message,
+								       session_descriptor->src_ae_inst_,
+								       session_descriptor->src_ae_name_,
+								       session_descriptor->src_ap_inst_,
+								       session_descriptor->src_ap_name_,
+								       invoke_id,
+								       remote_id);
 		} catch (rina::Exception &e) {
 			LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
 		}
@@ -1269,90 +457,38 @@ void EnrollmentTask::connect(int invoke_id,
 		return;
 	}
 
-	//3 Initiate the enrollment
-	try{
-		rina::FlowInformation flowInformation = irm_->getNMinus1FlowInformation(session_descriptor->port_id_);
-		EnrollerStateMachine * enrollmentStateMachine = (EnrollerStateMachine *) createEnrollmentStateMachine(
-				session_descriptor->get_destination_application_process_naming_info(),
-				session_descriptor->port_id_, false, flowInformation.difName);
-		enrollmentStateMachine->connect(invoke_id, session_descriptor);
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems: %s", e.what());
-
-		try {
-			rina::RemoteProcessId remote_id;
-			remote_id.port_id_ = session_descriptor->port_id_;
-
-			rib_daemon_->openApplicationConnectionResponse(rina::CDAPMessage::AUTH_NONE,
-					rina::AuthValue(), session_descriptor->dest_ae_inst_, session_descriptor->dest_ae_name_,
-					session_descriptor->dest_ap_inst_, session_descriptor->dest_ap_name_, -2,
-					std::string(e.what()), session_descriptor->src_ae_inst_, session_descriptor->src_ae_name_,
-					session_descriptor->src_ap_inst_, session_descriptor->src_ap_name_, invoke_id, remote_id);
-		} catch (rina::Exception &e) {
-			LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-		}
-
-		deallocateFlow(session_descriptor->port_id_);
-	}
+	//3 Delegate further processing to the policy
+	IPCPEnrollmentTaskPS * ipcp_ps = dynamic_cast<IPCPEnrollmentTaskPS *>(ps);
+	assert(ipcp_ps);
+	ipcp_ps->connect_received(invoke_id, session_descriptor);
 }
 
 void EnrollmentTask::connectResponse(int result, const std::string& result_reason,
-		rina::CDAPSessionDescriptor * session_descriptor) {
+		rina::CDAPSessionDescriptor * session_descriptor)
+{
 	LOG_IPCP_DBG("Received M_CONNECT_R cdapMessage from portId %d", session_descriptor->port_id_);
 
-	try{
-		EnrolleeStateMachine * stateMachine = (EnrolleeStateMachine*)
-				getEnrollmentStateMachine(session_descriptor, false);
-		stateMachine->connectResponse(result, result_reason);
-	}catch(rina::Exception &e){
-		//Error getting the enrollment state machine
-		LOG_IPCP_ERR("Problems getting enrollment state machine: %s", e.what());
-
-		try {
-			rina::RemoteProcessId remote_id;
-			remote_id.port_id_ = session_descriptor->port_id_;
-
-			rib_daemon_->closeApplicationConnection(remote_id, 0);
-		} catch (rina::Exception &e) {
-			LOG_IPCP_ERR("Problems closing application connection: %s", e.what());
-		}
-
-		deallocateFlow(session_descriptor->port_id_);
-	}
+	IPCPEnrollmentTaskPS * ipcp_ps = dynamic_cast<IPCPEnrollmentTaskPS *>(ps);
+	assert(ipcp_ps);
+	ipcp_ps->connect_response_received(result, result_reason, session_descriptor);
 }
 
-void EnrollmentTask::nMinusOneFlowAllocated(rina::NMinusOneFlowAllocatedEvent * flowEvent) {
+void EnrollmentTask::nMinusOneFlowAllocated(rina::NMinusOneFlowAllocatedEvent * flowEvent)
+{
 	rina::EnrollmentRequest * request =
 			port_ids_pending_to_be_allocated_.erase(flowEvent->handle_);
 
-	if (!request){
+	if (!request) {
 		return;
 	}
 
-	EnrolleeStateMachine * enrollmentStateMachine = 0;
-
-	//1 Tell the enrollment task to create a new Enrollment state machine
-	try{
-		enrollmentStateMachine = (EnrolleeStateMachine *) createEnrollmentStateMachine(
-				request->neighbor_->name_, flowEvent->flow_information_.portId, true,
-				flowEvent->flow_information_.difName);
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problem retrieving enrollment state machine: %s", e.what());
-		delete request;
-		return;
-	}
-
-	//2 Tell the enrollment state machine to initiate the enrollment
-	// (will require an M_CONNECT message and a port Id)
-	try{
-		enrollmentStateMachine->initiateEnrollment(
-				request, flowEvent->flow_information_.portId);
-	}catch(rina::Exception &e){
-		LOG_IPCP_ERR("Problems initiating enrollment: %s", e.what());
-	}
+	IPCPEnrollmentTaskPS * ipcp_ps = dynamic_cast<IPCPEnrollmentTaskPS *>(ps);
+	assert(ipcp_ps);
+	ipcp_ps->initiate_enrollment(*flowEvent, request);
 }
 
-void EnrollmentTask::nMinusOneFlowAllocationFailed(rina::NMinusOneFlowAllocationFailedEvent * event) {
+void EnrollmentTask::nMinusOneFlowAllocationFailed(rina::NMinusOneFlowAllocationFailedEvent * event)
+{
 	rina::EnrollmentRequest * request =
 			port_ids_pending_to_be_allocated_.erase(event->handle_);
 
@@ -1377,7 +513,8 @@ void EnrollmentTask::nMinusOneFlowAllocationFailed(rina::NMinusOneFlowAllocation
 }
 
 void EnrollmentTask::enrollmentFailed(const rina::ApplicationProcessNamingInformation& remotePeerNamingInfo,
-		int portId, const std::string& reason, bool enrollee, bool sendReleaseMessage) {
+		int portId, const std::string& reason, bool sendReleaseMessage)
+{
 	LOG_IPCP_ERR("An error happened during enrollment of remote IPC Process %s because of %s",
 			remotePeerNamingInfo.getEncodedString().c_str(), reason.c_str());
 
@@ -1405,94 +542,11 @@ void EnrollmentTask::enrollmentFailed(const rina::ApplicationProcessNamingInform
 	}
 
 	//3 In the case of the enrollee state machine, reply to the IPC Manager
-	if (enrollee) {
-		rina::EnrollmentRequest * request = ((EnrolleeStateMachine *) stateMachine)->enrollment_request_;
-		if (request) {
-			if (request->ipcm_initiated_) {
-				try {
-					rina::extendedIPCManager->enrollToDIFResponse(request->event_, -1,
-							std::list<rina::Neighbor>(), ipcp->get_dif_information());
-				} catch (rina::Exception &e) {
-					LOG_IPCP_ERR("Problems sending message to IPC Manager: %s", e.what());
-				}
-			}
-		}
-	}
+	IPCPEnrollmentTaskPS * ipcp_ps = dynamic_cast<IPCPEnrollmentTaskPS *>(ps);
+	assert(ipcp_ps);
+	ipcp_ps->inform_ipcm_about_failure(stateMachine);
 
 	delete stateMachine;
-}
-
-//Class EnrollmentRIBObject
-EnrollmentRIBObject::EnrollmentRIBObject(IPCProcess * ipc_process) :
-	BaseIPCPRIBObject(ipc_process, EncoderConstants::ENROLLMENT_INFO_OBJECT_CLASS,
-			rina::objectInstanceGenerator->getObjectInstance(), EncoderConstants::ENROLLMENT_INFO_OBJECT_NAME) {
-	enrollment_task_ = (EnrollmentTask *) ipc_process->enrollment_task_;
-	cdap_session_manager_ = ipc_process->cdap_session_manager_;
-}
-
-const void* EnrollmentRIBObject::get_value() const {
-	return 0;
-}
-
-void EnrollmentRIBObject::remoteStartObject(void * object_value, int invoke_id,
-		rina::CDAPSessionDescriptor * cdapSessionDescriptor) {
-	EnrollerStateMachine * stateMachine = 0;
-
-	try {
-		stateMachine = (EnrollerStateMachine *) enrollment_task_->getEnrollmentStateMachine(
-				cdapSessionDescriptor->dest_ap_name_, cdapSessionDescriptor->port_id_, false);
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems retrieving state machine: %s", e.what());
-		sendErrorMessage(cdapSessionDescriptor);
-		return;
-	}
-
-	if (!stateMachine) {
-		LOG_IPCP_ERR("Got a CDAP message that is not for me ");
-		return;
-	}
-
-	EnrollmentInformationRequest * eiRequest = 0;
-	if (object_value) {
-		eiRequest = (EnrollmentInformationRequest *) object_value;
-	}
-	stateMachine->start(eiRequest, invoke_id, cdapSessionDescriptor);
-}
-
-void EnrollmentRIBObject::remoteStopObject(void * object_value, int invoke_id,
-		rina::CDAPSessionDescriptor * cdapSessionDescriptor) {
-	EnrolleeStateMachine * stateMachine = 0;
-
-	try {
-		stateMachine = (EnrolleeStateMachine *) enrollment_task_->getEnrollmentStateMachine(
-				cdapSessionDescriptor->dest_ap_name_, cdapSessionDescriptor->port_id_, false);
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems retrieving state machine: %s", e.what());
-		sendErrorMessage(cdapSessionDescriptor);
-		return;
-	}
-
-	if (!stateMachine) {
-		LOG_IPCP_ERR("Got a CDAP message that is not for me");
-		return;
-	}
-
-	EnrollmentInformationRequest * eiRequest = 0;
-	if (object_value) {
-		eiRequest = (EnrollmentInformationRequest *) object_value;
-	}
-	stateMachine->stop(eiRequest, invoke_id , cdapSessionDescriptor);
-}
-
-void EnrollmentRIBObject::sendErrorMessage(const rina::CDAPSessionDescriptor * cdapSessionDescriptor) {
-	try{
-		rina::RemoteProcessId remote_id;
-		remote_id.port_id_ = cdapSessionDescriptor->port_id_;
-
-		rib_daemon_->closeApplicationConnection(remote_id, 0);
-	} catch (rina::Exception &e) {
-		LOG_IPCP_ERR("Problems sending CDAP message: %s", e.what());
-	}
 }
 
 // Class Operational Status RIB Object
