+++ rinad/src/ipcp/enrollment-task.h
@@ -33,21 +33,6 @@
 
 namespace rinad {
 
-/// The object that contains all the information
-/// that is required to initiate an enrollment
-/// request (send as the objectvalue of a CDAP M_START
-/// message, as specified by the Enrollment spec)
-class EnrollmentInformationRequest {
-public:
-	EnrollmentInformationRequest();
-
-	/// The address of the IPC Process that requests
-	///to join a DIF
-	unsigned int address_;
-	std::list<rina::ApplicationProcessNamingInformation> supporting_difs_;
-	bool allowed_to_start_early_;
-};
-
 class WatchdogRIBObject;
 
 /// Sends a CDAP message to check that the CDAP connection
@@ -101,173 +86,6 @@ private:
 	unsigned int address_;
 };
 
-/// The base class that contains the common aspects of both
-/// enrollment state machines: the enroller side and the enrolle
-/// side
-class BaseEnrollmentStateMachine : public rina::IEnrollmentStateMachine {
-public:
-	static const std::string CONNECT_RESPONSE_TIMEOUT;
-	static const std::string START_RESPONSE_TIMEOUT;
-	static const std::string START_IN_BAD_STATE;
-	static const std::string STOP_ENROLLMENT_TIMEOUT;
-	static const std::string STOP_IN_BAD_STATE;
-	static const std::string STOP_WITH_NO_OBJECT_VALUE;
-	static const std::string READ_RESPONSE_TIMEOUT;
-	static const std::string PROBLEMS_COMMITTING_ENROLLMENT_INFO;
-	static const std::string START_TIMEOUT;
-	static const std::string READ_RESPONSE_IN_BAD_STATE;
-	static const std::string UNSUCCESSFULL_READ_RESPONSE;
-	static const std::string UNSUCCESSFULL_START;
-	static const std::string CONNECT_IN_NOT_NULL;
-	static const std::string ENROLLMENT_NOT_ALLOWED;
-	static const std::string START_ENROLLMENT_TIMEOUT;
-	static const std::string STOP_ENROLLMENT_RESPONSE_TIMEOUT;
-	static const std::string STOP_RESPONSE_IN_BAD_STATE;
-
-	static const std::string STATE_WAIT_CONNECT_RESPONSE;
-	static const std::string STATE_WAIT_START_ENROLLMENT_RESPONSE;
-	static const std::string STATE_WAIT_READ_RESPONSE;
-	static const std::string STATE_WAIT_START;
-	static const std::string STATE_WAIT_START_ENROLLMENT;
-	static const std::string STATE_WAIT_STOP_ENROLLMENT_RESPONSE;
-
-	~BaseEnrollmentStateMachine(){};
-
-protected:
-	BaseEnrollmentStateMachine(IPCProcess * ipc_process,
-			const rina::ApplicationProcessNamingInformation& remote_naming_info,
-			int timeout, rina::ApplicationProcessNamingInformation * supporting_dif_name);
-
-	/// Sends all the DIF dynamic information
-	void sendDIFDynamicInformation();
-
-	/// Send the entries in the DFT (if any)
-	void sendDFTEntries();
-
-	IPCProcess * ipc_process_;
-	rina::CDAPSessionManagerInterface * cdap_session_manager_;
-	rina::IMasterEncoder * encoder_;
-};
-
-/// The state machine of the party that wants to
-/// become a new member of the DIF.
-class EnrolleeStateMachine: public BaseEnrollmentStateMachine {
-public:
-	EnrolleeStateMachine(IPCProcess * ipc_process,
-			const rina::ApplicationProcessNamingInformation& remote_naming_info,
-			int timeout);
-	~EnrolleeStateMachine();
-
-	/// Called by the DIFMembersSetObject to initiate the enrollment sequence
-	/// with a remote IPC Process
-	/// @param enrollment request contains information on the neighbor and on the
-	/// enrollment request event
-	/// @param portId
-	void initiateEnrollment(rina::EnrollmentRequest * enrollmentRequest, int portId);
-
-	/// Called by the EnrollmentTask when it got an M_CONNECT_R message
-	/// @param result
-	/// @param result_reason
-	void connectResponse(int result, const std::string& result_reason);
-
-	void startResponse(int result, const std::string& result_reason,
-			void * object_value, rina::CDAPSessionDescriptor * session_descriptor);
-
-	/// Stop enrollment request received. Check if I have enough information, if not
-	/// ask for more with M_READs.
-	/// Have to check if I can start operating (if not wait
-	/// until M_START operationStatus). If I can start and have enough information,
-	/// create or update all the objects received during the enrollment phase.
-	void stop(EnrollmentInformationRequest * eiRequest, int invoke_id,
-			rina::CDAPSessionDescriptor * cdapSessionDescriptor);
-
-	/// See if the response is valid and contains an object. See if more objects
-	/// are required. If not, start
-	void readResponse(int result, const std::string& result_reason,
-			void * object_value, const std::string& object_name,
-			rina::CDAPSessionDescriptor * session_descriptor);
-
-	void start(int result, const std::string& result_reason,
-			rina::CDAPSessionDescriptor * session_descriptor);
-
-	rina::EnrollmentRequest * enrollment_request_;
-
-private:
-	/// See if more information is required for enrollment, or if we can
-	/// start or if we have to wait for the start message
-	void requestMoreInformationOrStart();
-
-	/// Checks if more information is required for enrollment
-	/// (At least there must be DataTransferConstants, a QoS cube and a DAF Member). If there is,
-	/// it sends a read request message and returns tru
-	bool sendNextObjectRequired();
-
-	/// Create the objects in the RIB
-	void commitEnrollment();
-
-	void enrollmentCompleted();
-
-	bool was_dif_member_before_enrollment_;
-	bool allowed_to_start_early_;
-	int stop_request_invoke_id_;
-};
-
-/// The state machine of the party that is a member of the DIF
-/// and will help the joining party (enrollee) to join the DIF
-class EnrollerStateMachine: public BaseEnrollmentStateMachine {
-public:
-	EnrollerStateMachine(IPCProcess * ipc_process,
-		const rina::ApplicationProcessNamingInformation& remote_naming_info, int timeout,
-		rina::ApplicationProcessNamingInformation * supporting_dif_name);
-	~EnrollerStateMachine();
-
-	/// An M_CONNECT message has been received.  Handle the transition from the
-	/// NULL to the WAIT_START_ENROLLMENT state.
-	/// Authenticate the remote peer and issue a connect response
-	/// @param invoke_id
-    /// @param portId
-	void connect(int invoke_id, rina::CDAPSessionDescriptor * session_descriptor);
-
-	/// Called by the Enrollment object when it receives an M_START message from
-	/// the enrolling member. Have to look at the enrollment information request,
-	/// from that deduce if the IPC process requesting to enroll with me is already
-	/// a member of the DIF and if its address is valid. If it is not a member of
-	/// the DIF, send a new address with the M_START_R, send the M_CREATEs to provide
-	/// the DIF initialization information and state, and send M_STOP_R. If it is a
-	/// valid member, just send M_START_R with no address and M_STOP_R
-	/// @param eiRequest
-	/// @param invoke_id to reply to the message
-    /// @param cdapSessionDescriptor
-    void start(EnrollmentInformationRequest * eiRequest, int invoke_id,
-			rina::CDAPSessionDescriptor * cdapSessionDescriptor);
-
-    /// The response of the stop operation has been received, send M_START operation without
-    /// waiting for an answer and consider the process enrolled
-    /// @param result
-    /// @param result_reason
-    /// @param object_value
-    /// @param cdapSessionDescriptor
-	void stopResponse(int result, const std::string& result_reason,
-			void * object_value, rina::CDAPSessionDescriptor * session_descriptor);
-
-private:
-    /// Send a negative response to the M_START enrollment message
-    /// @param result the error code
-    /// @param resultReason the reason of the bad result
-    /// @param requestMessage the received M_START enrollment message
-    void sendNegativeStartResponseAndAbortEnrollment(int result, const std::string&
-    		resultReason, int invoke_id);
-
-    /// Send all the information required to start operation to
-    /// the IPC process that is enrolling to me
-    void sendDIFStaticInformation();
-
-    void enrollmentCompleted();
-
-	ISecurityManager * security_manager_;
-	INamespaceManager * namespace_manager_;
-};
-
 class EnrollmentTask: public IPCPEnrollmentTask {
 public:
 	EnrollmentTask();
@@ -281,20 +99,10 @@ public:
 	void connectResponse(int result, const std::string& result_reason,
 			rina::CDAPSessionDescriptor * session_descriptor);
 	void enrollmentFailed(const rina::ApplicationProcessNamingInformation& remotePeerNamingInfo,
-			int portId, const std::string& reason, bool enrolle, bool sendReleaseMessage);
+			int portId, const std::string& reason, bool sendReleaseMessage);
 
 private:
 	void populateRIB();
-
-	/// Creates an enrollment state machine with the remote IPC process identified by the apNamingInfo
-	/// @param apNamingInfo
-	/// @param enrollee true if this IPC process is the one that initiated the
-	/// enrollment sequence (i.e. it is the application process that wants to
-	/// join the DIF)
-	/// @return
-	rina::IEnrollmentStateMachine * createEnrollmentStateMachine(
-			const rina::ApplicationProcessNamingInformation& apNamingInfo, int portId,
-			bool enrollee, const rina::ApplicationProcessNamingInformation& supportingDifName);
 
 	/// Called when a new N-1 flow has been allocated
 	// @param portId
@@ -308,23 +116,6 @@ private:
 
 	INamespaceManager * namespace_manager_;
 	rina::Thread * neighbors_enroller_;
-};
-
-/// Handles the operations related to the "daf.management.enrollment" objects
-class EnrollmentRIBObject: public BaseIPCPRIBObject {
-public:
-	EnrollmentRIBObject(IPCProcess * ipc_process);
-	const void* get_value() const;
-	void remoteStartObject(void * object_value, int invoke_id,
-			rina::CDAPSessionDescriptor * cdapSessionDescriptor);
-	void remoteStopObject(void * object_value, int invoke_id,
-			rina::CDAPSessionDescriptor * cdapSessionDescriptor);
-
-private:
-	void sendErrorMessage(const rina::CDAPSessionDescriptor * cdapSessionDescriptor);
-
-	EnrollmentTask * enrollment_task_;
-	rina::CDAPSessionManagerInterface * cdap_session_manager_;
 };
 
 /// Handles the operations related to the "daf.management.operationalStatus" object
