+++ rinad/src/ipcp/flow-allocator.h
@@ -180,10 +180,6 @@ public:
 	IFlowAllocatorInstance * getFAI(int portId);
 	void set_application_process(rina::ApplicationProcess * ap);
 	void set_dif_configuration(const rina::DIFConfiguration& dif_configuration);
-        int select_policy_set(const std::string& path, const std::string& name);
-        int set_policy_set_param(const std::string& path,
-                                 const std::string& name,
-                                 const std::string& value);
 	void createFlowRequestMessageReceived(Flow * flow, const std::string& object_name,
 			int invoke_id, int underlying_port_id);
 	void submitAllocateRequest(rina::FlowRequestEvent& flowRequestEvent);
