+++ rinad/src/ipcp/namespace-manager.cc
@@ -612,17 +612,4 @@ unsigned int NamespaceManager::getAdressByname(const rina::ApplicationProcessNam
 	throw rina::Exception("Unknown neighbor");
 }
 
-int NamespaceManager::select_policy_set(const std::string& path,
-                                     const std::string& name)
-{
-	return select_policy_set_common(get_name(), path, name);
-}
-
-int NamespaceManager::set_policy_set_param(const std::string& path,
-                                        const std::string& name,
-                                        const std::string& value)
-{
-	return set_policy_set_param_common(path, name, value);
-}
-
 }
