+++ rinad/src/ipcp/resource-allocator.h
@@ -61,10 +61,6 @@ public:
 	void set_application_process(rina::ApplicationProcess * ap);
 	void set_dif_configuration(const rina::DIFConfiguration& dif_configuration);
 	INMinusOneFlowManager * get_n_minus_one_flow_manager() const;
-	int select_policy_set(const std::string& path, const std::string& name);
-	int set_policy_set_param(const std::string& path,
-			const std::string& name,
-			const std::string& value);
 
 private:
 	INMinusOneFlowManager * n_minus_one_flow_manager_;
